refactor(core): refactor OOP implementation
